fix: Claude forgot to change an int into a Long

+++ element-template-generator/core/src/main/java/io/camunda/connector/generator/api/CliCompatibleTemplateGenerator.java
@@ -47,7 +47,7 @@ public interface CliCompatibleTemplateGenerator<IN> extends ElementTemplateGener
   record ScanResult(
       String templateId,
       String templateName,
-      Integer templateVersion,
+      Long templateVersion,
       String templateType,
       Object additionalData) {}
 }
